// src/app/models/formsStep.model.ts
export interface FormsStepsInterface {
    currentProductStep: string;
    currentServiceStep: string;
    currentRegistrationStep: string;
    currentServiceProgress: number;
    currentProductProgress: number;
}

export class FormsStepsModel {
    currentProductStep: string;
    currentServiceStep: string;
    currentRegistrationStep: string;
    currentServiceProgress: number;
    currentProductProgress: number;
}